pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "correttore"
version = "2.0.0"
description = "Enterprise Italian Text Correction System"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Marco LP", email = "your.email@example.com"}
]
keywords = ["nlp", "italian", "correction", "grammar", "spell-check"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Text Processing :: Linguistic",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "openai>=1.0.0",
    "python-docx>=0.8.11",
    "pyyaml>=6.0",
    "flask>=2.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "mypy>=0.990",
    "flake8>=5.0",
]

[project.scripts]
correttore = "correttore.interfaces.cli:main"
correttore-analyze = "correttore.utils.readability:main_cli"

[project.urls]
Homepage = "https://github.com/MarcoLP1822/correttore"
Documentation = "https://github.com/MarcoLP1822/correttore/docs"
Repository = "https://github.com/MarcoLP1822/correttore"
Issues = "https://github.com/MarcoLP1822/correttore/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=correttore --cov-report=html --cov-report=term"
